First Upload
Also attempting to create a collision that will load us into another
scene.

Running into an issue where the collision just doesn’t happen with the
box. It may be due to the First Person Controller or it maybe something
to do with the meshes??????

I DONT KNOW. Need to figure out.